[TASK] restructure namespace
